CI: us "$runs_on_pool", not "$jobname" to select packages & config
Change the setup hooks for the CI to use "$runs_on_pool" for the
"$regular" job. Now instead of needing to enumerate each jobname when
we select packages we can install things depending on the pool we're
running in. This is possible since we don't need to worry about Travis
CI anymore.

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>