CI: run "documentation" via run-build-and-test.sh
Now that we've taught the build and test script to accept arbitrary
"make" targets we can remove the "ci/test-documentation.sh" script in
favor of testing both:

    make check-builtins check-docs
    make check-builtins check-docs USE_ASCIIDOCTOR=Y

Note that this isn't exactly the same thing as the previous CI target,
as we're not asserting that we don't have output on STDERR, but we can
assume basic sanity in those targets, i.e. that they'll carry forward
an exit code, and check STDERR if it's relevant.

TODO: Fix up the "test -n" on the "gem install" thing into Travis
removal? See b011fabd6e0 (ci/lib.sh: encapsulate Travis-specific
things, 2019-01-27).

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>