CI: add and use a fake t/.prove
Generate a fake t/.prove where we pretend that the runtime of a given
test is how many "test_expect" etc. lines we can find in it. This is a
fuzzy heuristic for their runtime which mostly works out. See
b819f1d2cec (ci: parallelize testing on Windows, 2019-01-29) for an
alternate way of doing this with "stat".

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>