test-lib.sh: remove the test_untraceable facility
The last user of the "test_untraceable" facility added in
5fc98e79fc0 (t: add means to disable '-x' tracing for individual test
scripts, 2018-02-24) went away in the preceding commit.

It seems to me that for tests such as the now-amended
"t1510-repo-setup.sh" we're more likely to want to selectively skip
certain tests under tracing than to disable "-x" entirely. We can
always bring this facility back later if needed.

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>