setup: fall back to $PWD if getcwd() fails
Try a bit harder in setup.c to not error out on cases such as:

    $ (mkdir x && cd x && rmdir ../x && git status)
    fatal: Unable to read current working directory: No such file or directory

When we are running under a POSIX shell and the user has navigated to
the now-removed directory with "cd" we can can check what the
"current" directory is with "$PWD". That directory is of course gone,
but knowing what used to be the path is enough for us to start looking
for the .git directory. Now we'll gracefully recover instead:

    $ (mkdir x && cd x && rmdir ../x && ~/g/git/git status)
    On branch avar/setup-handle-gone-directory
    Your branch is ahead of 'origin/master' by 2 commits.
      (use "git push" to publish your local commits)

    Changes to be committed:
      (use "git restore --staged <file>..." to unstage)
            modified:   ../setup.c
            modified:   ../strbuf.c
            modified:   ../strbuf.h
            modified:   ../t/t1510-repo-setup.sh
            modified:   ../wrap-for-bin.sh
    [...]

We'll not only display the "status", but our relative paths (to the
now-disappeared directory) are correct.

There is a parallel proposed effort to change git itself to stop
removing the user's current directory[2], this change is orthogonal to
that as the user might end up with a removed directory due to using
something other than git to change the filesystem, so setup.c should
try to handle these cases.

This still needs some more work, some git commands work, such as
"status", it even works to do "git add ../strbuf.c" in this state, but
e.g. a subsequent "git reset ../strbuf.c" will error out.

The reason for that is that in check_filename() in setup.c we'll
construct an absolute path from the root on the basis of the prefix,
i.e. "../strbuf.c" will be converted to "x/../strbuf.c", and we'll
feed that to lstat(), which will of course fail as "x/" doesn't
exist.

But I don't see why we couldn't canonicalize that "x/../" to simply
"", and lstat("strbuf.c") from the root, which would make "reset"
work, and probably saves the C library/FS/kernel some slight bit of
work in either case (but I may be missing some edge case).

1. https://pubs.opengroup.org/onlinepubs/009604599/utilities/cd.html
2. https://lore.kernel.org/git/pull.1140.git.git.1637455620.gitgitgadget@gmail.com/

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>