linear-assignment.c: add and use intprops.h from Gnulib
In the preceding commit we started optimistically using
__builtin_{add,mul}_overflow under GCC and Clang, but as reading
through the intprops.h being added here from gnulib reveals, using
just those functions is a rabbit hole of version-specific GCC and
Clang bugs, compiler detection etc.

Let's outsource all that detection to the macro-only intprops.h from
Gnulib. This header is licensed under the LGPL 2.1, the same license
that xdiff/* and compat/regex/* are under. So there's no license
reason for why we can't import this.

This is the latest version from
gnulib.git (https://git.savannah.gnu.org/git/gnulib.git), last
modified with 5d88f13109 (regex: sync with glibc, 2021-09-21). The
once change to it is the removal of:

    #include <limits.h>

Which we'll get from git-compat-util.h.

Now we'll still die the same way on overflow, i.e:

    git -P range-diff --creation-factor=50 origin/master...git-for-windows/main
    fatal: integer overflow in cost[47395 + ((47396 * 45309) = -2147454537)] addition

But this time we don't need compiler-specific checks in our main
sources to do so, this even works on e.g. Intel ICC.

Note that this file has various "git diff --check" issues, hence the
"compat/gnulib/.gitattributes" being added here. These are external
sources, we should import them as close to as-is as we can, to make it
easier to upgrade them in the future.

We could go a step further here and replace the various
unsigned_*_overflows() in "git-compat-util.h" with macros from this
intprops.h header. See [1][2][3][4][5] for our home-grown
implementation of overflow checking. That could be done as simply
replacing them with:

    #include "compat/gnulib/intprops.h"
    #define signed_add_overflows(a, b) INT_ADD_OVERFLOW(a, b)
    #define unsigned_add_overflows(a, b) INT_ADD_OVERFLOW(a, b)
    #define unsigned_mult_overflows(a, b) INT_MULTIPLY_OVERFLOW(a, b)
    #define unsigned_left_shift_overflows(a, shift) INT_LEFT_SHIFT_OVERFLOW(a, shift)

The benefit of doing so is that we'd end up with implementations that
would be more robust, as they'd detect both signed and unsigned
overflows (we should then also rename our own macros).

The disadvantage is that it's a bit slower. I forward-ported the
benchmark program [6] written in 2016. Using the above macro
definitions the slowdown measured with git-hyperfine[7] is the
following:

    $ git hyperfine -L rev HEAD~1,HEAD~0 -s 'make CFLAGS=-O3 git-strbuf-grow' './git-strbuf-grow'
    Benchmark 1: ./git-strbuf-grow' in 'HEAD~1
      Time (mean ± σ):      2.661 s ±  0.025 s    [User: 2.657 s, System: 0.004 s]
      Range (min … max):    2.626 s …  2.713 s    10 runs

    Benchmark 2: ./git-strbuf-grow' in 'HEAD~0
      Time (mean ± σ):      2.875 s ±  0.031 s    [User: 2.871 s, System: 0.004 s]
      Range (min … max):    2.832 s …  2.948 s    10 runs

    Summary
      './git-strbuf-grow' in 'HEAD~1' ran
        1.08 ± 0.02 times faster than './git-strbuf-grow' in 'HEAD~0'

The slowdown is not overhead in Gnulib's intprops.h, but happens under
both GCC and Clang if we use __builtin_add_overflow() itself, so the
difference is in the infinite precision addition the compilers need to
do. We should probably not care and use them anyway, but let's just
narrowly fix the signed overflow issue for now.

1. c03c83152d6 (do not depend on signed integer overflow, 2010-10-05)
2. 1368f65002b (compat: helper for detecting unsigned overflow, 2010-10-10)
3. 320d0b493a2 (add helpers for detecting size_t overflow, 2016-02-19)
4. 935de81289c (add helpers for detecting size_t overflow, 2016-02-19)
5. e2ffeae3f67 (git-compat-util: introduce more size_t helpers, 2021-11-02)
6. https://lore.kernel.org/git/20160601210713.GA18118@sigill.intra.peff.net/
7. https://github.com/avar/git-hyperfine/

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>