tests: run the tests with GIT_NO_TEMPLATE_DIR=true
Add a GIT_NO_TEMPLATE_DIR environment variable which overrides the
"GIT_TEMPLATE_DIR" set in both t/test-lib.sh and
wrap-for-bin.sh.

Using it makes any "git [init|clone|submodule|worktree]" command that
initializes a repository use the the equivalent of a "--no-template"
option (which isn't available for "submodule" and "worktree"), unless
explicitly overridden by config or CLI option.

This change reduces the I/O the test suite generates by quite a bit,
and makes debugging easier by reducing the set of files irrelevant to
a failing test. In the case of submodules in particular we'd end up
with N copies of our template files for the N submodules (plus the
main repository).

Before this running it with --debug (so that trash is retained)
results in ~770MB of trash* (according to "du -shc"), after it's
reduced to around 590M.

That saving shows that almost 1/4 bytes we wrote were the same sample
hooks and other repetitive data. The number of files created went down
from ~151k to ~110k, which around the same reduction of (more than)
1/4.

This will also make our tests more reliable as we're now forced to
check whether our software works with any arbitrary --template that
may be in use in the wild, as opposed to our relatively "fat" current
default. Having our tests work this way would have caught the issue
with "git sparse-checkout" fixed in 7f44842ac19 (sparse-checkout:
create leading directory, 2022-01-21).

This change benefits from preceding commits which fixed implicit
dependencies on our templates. See subjects with titles starting with
"tests: don't assume a " and "tests: don't depend on ".

Because of those changes and the introduction of the "test_hook"
helper we can now trivially change just "test_hook" to make the
".git/hook" directory, and to "rmdir" it during cleanup. As a TODO
comment notes we might be even stricter about this in the future, but
for now we those up lazily, even if we're strict about setting up and
cleaning up hooks with --setup and --clobber.

One exception to being able to use "test_hook" is a test in
"t7450-bad-git-dotfiles.sh" which sets up an intentionally broken
submodule. We cannot use "git rev-parse" to discover the canonical
".git" dir of that broken submodule, and therefore cannot use
"test_hook". Let's instead do the equivalent of a manual "mkdir
.git/hooks" in that case.

Similarly a test in "t7814-grep-recurse-submodules.sh" needed to be
manually adjusted, but only because "git submodule" was not paying
attention to --template= or "TEST_CREATE_REPO_NO_TEMPLATE".

With this change we'll also be able to remove the transitory
"TEST_CREATE_REPO_NO_TEMPLATE" variable, and instances of invoking
"git [init|clone]" with "--template=". That cleanup will be done in a
subsequent commit.

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>