test-lib: create test data with "git init --no-template" (almost)
Change "t/test-lib.sh" to set the newly added
"GIT_NO_TEMPLATE_DIR=true" option in the environment, which makes any
"git init" invocation take the equivalent of a "--no-template" option,
unless explicitly overridden by config or CLI option.

Well, "almost". We're also adding a test-only
"GIT_TEST_BARE_TEMPLATE=true", when it's set we'll interpret
"GIT_NO_TEMPLATE_DIR=true" as meaning no template, except for empty
"info", "hooks" and "branches" directories.

This is because we've added various implicit dependencies on having
these created by "git init". In subsequent commits we'll address some
of those. More on that below.

This reduces the I/O the test suite generates by quite a bit. Before
this running it with --debug (so that trash is retained) results in
~770MB of trash* (according to "du -shc"), after it's reduced to
around 590M, so almost 1/4 bytes we wrote were the same sample hooks
and other repetitive data. The number of files created went down from
~151k to ~110k, which around the same reduction of (more than) 1/4.

This will also make our tests more reliable as we're now forced to
check whether our software works with any arbitrary --template that
may be in use in the wild, as opposed to our relatively "fat" current
default.

On the "more on that below", these are the directories we're
creating (only in the test suite) for now:

- "hooks": Removing it will be relatively simple, mostly a matter of a
  few tests needing a "mkdir .git/hooks", or the tests assume that
  they'll need to explicitly disable .git/hooks.

- "branches": Similarly trivial. The last attempt to get rid of it was
  aborted in c8a58ac5a52 (Revert "Don't create the $GIT_DIR/branches
  directory on init", 2009-10-31), but only a couple of tests files
  depend on it.

- "info": This is the big one. Mostly these are tests that expect to
  echo to .git/info/excludes.

  But we've also got "git sparse-checkout" hard-depending on its
  presence since 94c0956b609 (sparse-checkout: create builtin with
  'list' subcommand, 2019-11-21) and bab3c359087 (sparse-checkout:
  create 'init' subcommand, 2019-11-21). I.e. we'd hard error if the
  template hadn't created this directory.

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>